requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Science Essentials
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Image Processing
Pillow>=9.5.0

# Progress Bars
tqdm>=4.65.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.22.0

# Web App Demo
streamlit>=1.22.0

# Model Evaluation
scikit-learn>=1.2.0
